+++ authority-portal-frontend/src/app/pages/authority-organization-detail-page/authority-organization-detail-page/authority-organization-detail-page.component.ts
@@ -5,7 +5,7 @@ import {Store} from '@ngxs/store';
 import {DeploymentEnvironmentDto} from '@sovity.de/authority-portal-client';
 import {GlobalStateUtils} from 'src/app/core/global-state/global-state-utils';
 import {mapRolesToReadableFormat} from 'src/app/core/utils/user-role-utils';
-import {OrganizationActions} from 'src/app/shared/components/organization-detail/organization-detail.component';
+import {OrganizationActions} from 'src/app/shared/components/business/organization-detail/organization-detail.component';
 import {
   ApproveOrganization,
   RefreshOrganization,