+++ authority-portal-frontend/src/app/pages/participant-own-connector-list-page/participant-own-connector-list-page/participant-own-connector-list-page.model.ts
@@ -1,0 +1,4 @@
+export enum ParticipantOwnConnectorListHeaderActions {
+  ADD_CONNECTOR = 'ADD_CONNECTOR',
+  PROVIDE_CONNECTOR = 'PROVIDE_CONNECTOR',
+}