+++ authority-portal-frontend/src/app/shared/components/common/header-bar/header-bar.component.ts
@@ -1,0 +1,15 @@
+import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {HeaderBarConfig} from 'src/app/shared/components/common/header-bar/header-bar.model';
+
+@Component({
+  selector: 'app-header',
+  templateUrl: './header-bar.component.html',
+})
+export class HeaderBarComponent {
+  @Input() headerConfig!: HeaderBarConfig;
+  @Output() headerActionClicked = new EventEmitter<string>();
+
+  onHeaderActionClicked(action: string): void {
+    this.headerActionClicked.emit(action);
+  }
+}